Temp rename of temperature widget
